Application Pre-Alpha Re-upload
Re-upload of the Pre-Alpha version of the application with minor changes.

In the "Distractors With Two Tables" experiment type, it is now impossible for any of the distractors to have the same shape and colour as the target object unless there are no other alternatives.

Additionally, the user's rotation can also be disabled under the same condition in the configuration file.